Changes on flavors
Old flavors has been changed to "beta" and "stable", so beta versions of
husky can be released with the proper flavor config, along the name of
the application, the package name, and the versions.

In the future it will ship ACRA too, so random crashes can be reported
with an email (using a template).