New flavors: dev and newhusky
As I explained earlier when I took over Husky maintenance, I'm making
changes in the application to add stuff that would make it more
maintenable in the long term, that is, an architecture, new stuff that
would make the application better in performance, while also keeping all
the good stuff of Husky.

The flavor 'newhusky' is for that matter. I'm going to keep working on
keeping Husky maintained, and at the same time I'll make the "new" Husky
in another flavor. It won't change anything because, being two flavors,
releases will be made with the current application, so be patient,
please.