NotificationsAdapter: implement custom emoji reacts
If the emoji_url field is specified in the notification data, it will be
used to represent the emoji in the notification text. As such, the
Notification entity data class has been updated accordingly, to match
the current implementation at AkkomaDev/Akkoma, which may later be
integrated by more Pleroma-based backends.

The SpannableBuilder formatting is a bit hacky, but it works. Ideally,
something like Phrase should be used for localization and Spannable
formatting. However, this would be a different undertaking of its own,
and would require more work on refactoring multiple parts of the
application code.

Signed-off-by: Hélène <pleroma-dev@helene.moe>
Signed-off-by: Adolfo Santiago <epoch@nixnetmail.com>