Create method to get settings from DB
The InstanceRepository provides a method to get the data from the DB.
This is because, in case the endpoint fails, the error will be catched
and, instead of showing an error, the user will be able to keep editing
their profile because the settings are stored.

If the api/v1/instance endpoint fails, there is no reason that the
profile endpoint will fail, thus the user will be able to edit their
profile.